Added North and East step calculation that multiplies each step by the sin and cos of the bearing in radians. The components of each step are updated constantly with respect to the current bearing and not the bearing when the step is counted.

+++ app/src/main/java/ca/uwaterloo/lab3_201_04/accValues.java
@@ -16,8 +16,8 @@ public class accValues {
 
     static int stepCount = 0;
     static boolean stepCheckEnabled = true;
-    static float stepCountNorth = 0; // TODO: Use the stepCountNorth/East variables.
-    static float stepCountEast = 0;
+    static double stepCountNorth = 0; // TODO: Use the stepCountNorth/East variables.
+    static double stepCountEast = 0;
 
     // Min, max, and reset amplitudes for the algorithm in m/s^2.
     static double minAmplitude = 0.5;
